+++ scripts/get_refresh_token.py
@@ -14,6 +14,7 @@ import os
 
 SCOPES = [
     'https://www.googleapis.com/auth/gmail.compose',
+    'https://www.googleapis.com/auth/gmail.send',
     'https://www.googleapis.com/auth/drive',
     'https://www.googleapis.com/auth/calendar',
 ]